refactor: [US-059] - Move recommendations + maintenance tasks; thin core/tasks/__init__ aggregator
